requirements.txt
pandas==2.2.2
pathlib==1.0.1
scikit-learn==1.4.2
xgboost==2.0.3
psycopg2-binary
